Clean EslDbContext in Infrastructure project with update Business Entities in Core project.  Next is EslViewContext for naming convention

+++ Core/Models/BusinessEntities/RptAllEvent.cs
@@ -8,7 +8,7 @@ public partial record RptAllEvent
 
     public string? EventID { get; set; }
 
-    public byte? EventID_RevNo { get; set; }
+    public int? EventID_RevNo { get; set; }
 
     public DateTime? EventDate { get; set; }
